third book commit

+++ src/main/java/com/digit/ecommerce/service/BookService.java
@@ -4,7 +4,6 @@ package com.digit.ecommerce.service;
 
 import com.digit.ecommerce.dto.BooksDto;
 import com.digit.ecommerce.dto.DataHolder;
-import com.digit.ecommerce.exception.BookAlreadyExistsException;
 import com.digit.ecommerce.exception.DataNotFoundException;
 import com.digit.ecommerce.exception.RoleNotAllowedException;
 import com.digit.ecommerce.model.Books;
@@ -12,20 +11,19 @@ import com.digit.ecommerce.model.User;
 import com.digit.ecommerce.repository.BookRepository;
 import com.digit.ecommerce.repository.UserRepository;
 import com.digit.ecommerce.util.TokenUtility;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataRetrievalFailureException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
+
+import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
+
 
 
 @Service
-public class BookService {
+public class BookService implements BooksInterface{
 
     @Autowired
     BookRepository bookRepository;
@@ -37,15 +35,14 @@ public class BookService {
     UserRepository userRepository;
 
 
-
-    public ResponseEntity<?> addBooks(BooksDto booksDto, String token) throws RuntimeException {
+    public ResponseEntity<?> addBooks(BooksDto booksDto, String token) {
         DataHolder dataHolder = tokenUtility.decode(token);
         String requiredRole = "admin";
         Long Admin_id = dataHolder.getId();
         Books existBook = null;
-        User objUser=userRepository.findById(Admin_id).orElse(null);
-        Books objBook=new Books(booksDto);
-        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())&& token!=null && objUser!=null) {
+        User objUser = userRepository.findById(Admin_id).orElse(null);
+        Books objBook = new Books(booksDto);
+        if (requiredRole.equalsIgnoreCase(dataHolder.getRole()) && token != null && objUser != null) {
             existBook = bookRepository.findBybookName(booksDto.getBookName());
             if (existBook == null) {
                 Books savedBook = bookRepository.save(objBook);
@@ -64,40 +61,45 @@ public class BookService {
     }
 
 
-
     public List<BooksDto> viewAllBooks(String token) {
         DataHolder dataHolder = tokenUtility.decode(token);
-        if ((dataHolder.getRole().equalsIgnoreCase("Admin") || dataHolder.getRole().equalsIgnoreCase("User"))&& token!=null){
+        Long Admin_id = dataHolder.getId();
+        User objUser = userRepository.findById(Admin_id).orElse(null);
+        if ((dataHolder.getRole().equalsIgnoreCase("Admin") || dataHolder.getRole().equalsIgnoreCase("User")) && token != null && objUser != null) {
             List<Books> booksList = bookRepository.findAll();
             List<BooksDto> booksDtoList = booksList.stream()
                     .map(BooksDto::new)
                     .toList();
             return booksDtoList;
-        }
-        else{
+        } else {
             throw new RoleNotAllowedException("Please login to Access the data");
         }
     }
-    public String deleteBook(Long id,String token) {
+
+    public String deleteBook(Long id, String token) {
         DataHolder dataHolder = tokenUtility.decode(token);
         String requiredRole = "admin";
-        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())) {
+        Long Admin_id = dataHolder.getId();
+        User objUser = userRepository.findById(Admin_id).orElse(null);
+        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())&& objUser != null) {
             if (!bookRepository.existsById(id)) {
                 throw new DataNotFoundException("Book with ID " + id + " not found");
             }
             bookRepository.deleteById(id);
             return "Deleted successfully";
-        }
-        else {
+        } else {
             throw new RoleNotAllowedException("Only admin can have Access to Delete");
         }
     }
 
 
-    public BooksDto updateBooks(Long id, Books bookDetails , String token) {
+    public BooksDto updateBooks(Long id, BooksDto bookDetails , String token) {
         DataHolder dataHolder = tokenUtility.decode(token);
         String requiredRole = "admin";
-        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())) {
+        Books books=new Books(bookDetails);
+        Long Admin_id = dataHolder.getId();
+        User objUser = userRepository.findById(Admin_id).orElse(null);
+        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())&& objUser != null) {
             Books existing = bookRepository.findById(id)
                     .orElseThrow(() -> new RuntimeException("Book not found with id: " + id));
             if (bookDetails.getBookAuthor() != null) {
@@ -106,9 +108,6 @@ public class BookService {
             if (bookDetails.getBookDescription() != null) {
                 existing.setBookDescription(bookDetails.getBookDescription());
             }
-//            if (bookDetails.getBookLogo() != null) {
-//                existing.setBookLogo(bookDetails.getBookLogo());
-//            }
             if (bookDetails.getBookPrice() != null) {
                 existing.setBookPrice(bookDetails.getBookPrice());
             }
@@ -127,10 +126,13 @@ public class BookService {
     }
 
 
-    public BooksDto updatePrice( Long id, Books books, String token) throws RuntimeException {
+    public BooksDto updatePrice( Long id, BooksDto booksDto, String token) throws RuntimeException {
         DataHolder dataHolder = tokenUtility.decode(token);
+        Books books=new Books(booksDto);
         String requiredRole = "admin";
-        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())) {
+        Long Admin_id = dataHolder.getId();
+        User objUser = userRepository.findById(Admin_id).orElse(null);
+        if (requiredRole.equalsIgnoreCase(dataHolder.getRole())&& objUser != null) {
             Books existing = bookRepository.findById(id)
                     .orElseThrow(() -> new RuntimeException("Book not found with id " + id));
             if (books.getBookPrice() != null) {
@@ -144,5 +146,21 @@ public class BookService {
         }
     }
 
+//public void updateQuantity(Order order) {
+    // Order BookId = bookRepository.getById();
+//        Books books = bookRepository.findById(book_Id).orElse(null);
+//        if (books != null) {
+//            if (status.equalsIgnoreCase("ordered")) {
+//                books.setBookQuantity(books.getBookQuantity() - quantity);
+//            } else {
+//                books.setBookQuantity(books.getBookQuantity() + quantity);
+//            }
+//            bookRepository.save(books);
+//        } else {
+//            // Handle the case where the book is not found
+//            System.out.println("Book not found");
+//        }
+//    }
 }
+
